backend/requirements.txt
Django>=4.2
gunicorn
whitenoise
django-cors-headers
mysqlclient
python-dotenv


